+++ CHANGELOG.md
@@ -2,7 +2,8 @@
 
 ## Unreleased
 
-- Error handling when parsing outdated stdout.
+- Show STDERR from outdated checks in progress bar.
+- Error handling when parsing outdated STDOUT.
 
 ## 2.1.1